+++ Zenith-LanguageModel/src/main/java/com/ciphertool/zenith/model/dao/LetterNGramDao.java
@@ -95,23 +95,23 @@ public class LetterNGramDao {
 			while (cursor.hasNext()) {
 				next = cursor.next();
 
-				if (Double.parseDouble((String) next.get(COUNT_KEY)) > (double) minimumCount) {
+				if (Float.parseFloat((String) next.get(COUNT_KEY)) > (float) minimumCount) {
 					nextNode = new TreeNGram((String) next.get(CUMULATIVE_STRING_KEY));
 
 					nextNode.setId((ObjectId) next.get(ID_KEY));
-					nextNode.setCount(Double.parseDouble((String) next.get(COUNT_KEY)));
+					nextNode.setCount(Float.parseFloat((String) next.get(COUNT_KEY)));
 					nextNode.setOrder((int) next.get(ORDER_KEY));
 
 					if (next.containsField(PROBABILITY_KEY)) {
-						nextNode.setProbability(Double.parseDouble((String) next.get(PROBABILITY_KEY)));
+						nextNode.setProbability(Float.parseFloat((String) next.get(PROBABILITY_KEY)));
 					}
 
 					if (next.containsField(CONDITIONAL_PROBABILITY_KEY)) {
-						nextNode.setConditionalProbability(Double.parseDouble((String) next.get(CONDITIONAL_PROBABILITY_KEY)));
+						nextNode.setConditionalProbability(Float.parseFloat((String) next.get(CONDITIONAL_PROBABILITY_KEY)));
 					}
 
 					if (next.containsField(CHAINED_PROBABILITY_KEY)) {
-						nextNode.setChainedProbability(Double.parseDouble((String) next.get(CHAINED_PROBABILITY_KEY)));
+						nextNode.setChainedProbability(Float.parseFloat((String) next.get(CHAINED_PROBABILITY_KEY)));
 					}
 
 					nodesToReturn.add(nextNode);
